+++ lang_verilog.cc
@@ -108,7 +108,7 @@ static void parse_args_paramset(CS& cmd, MODEL_CARD* x)
   assert(x);
 
   while (cmd >> '.') {
-    unsigned here = cmd.cursor();
+    size_t here = cmd.cursor();
     std::string name, value;
     try{
       cmd >> name >> '=' >> value >> ';';
@@ -127,7 +127,7 @@ static void parse_args_instance(CS& cmd, CARD* x)
     if (cmd.match1('.')) {
       // by name
       while (cmd >> '.') {
-	unsigned here = cmd.cursor();
+	size_t here = cmd.cursor();
 	std::string name  = cmd.ctos("(", "", "");
 	std::string value = cmd.ctos(",)", "(", ")");
 	cmd >> ',';
@@ -141,7 +141,7 @@ static void parse_args_instance(CS& cmd, CARD* x)
       // by order
       int index = 1;
       while (cmd.is_alnum() || cmd.match1("+-.")) { untested();
-	unsigned here = cmd.cursor();
+	size_t here = cmd.cursor();
 	try{ untested();
 	  std::string value = cmd.ctos(",)", "", "");
 	  x->set_param_by_index(x->param_count() - index++, value, 0/*offset*/);
@@ -177,7 +177,7 @@ static void parse_ports(CS& cmd, COMPONENT* x, bool all_new)
       // by order
       int index = 0;
       while (cmd.is_alnum()) {
-	unsigned here = cmd.cursor();
+	size_t here = cmd.cursor();
 	try{
 	  std::string value;
 	  cmd >> value;
@@ -207,7 +207,7 @@ static void parse_ports(CS& cmd, COMPONENT* x, bool all_new)
     }else{ untested();
       // by name
       while (cmd >> '.') { untested();
-	unsigned here = cmd.cursor();
+	size_t here = cmd.cursor();
 	try{ untested();
 	  std::string name, value;
 	  cmd >> name >> '(' >> value >> ')' >> ',';
@@ -349,7 +349,7 @@ COMPONENT* LANG_VERILOG::parse_instance(CS& cmd, COMPONENT* x)
 /*--------------------------------------------------------------------------*/
 std::string LANG_VERILOG::find_type_in_string(CS& cmd)
 {
-  unsigned here = cmd.cursor();
+  size_t here = cmd.cursor();
   std::string type;
   if ((cmd >> "//")) {
     assert(here == 0);
@@ -519,7 +519,7 @@ class CMD_PARAMSET : public CMD {
     // already got "paramset"
     std::string my_name, base_name;
     cmd >> my_name;
-    unsigned here = cmd.cursor();    
+    size_t here = cmd.cursor();
     cmd >> base_name;
 
     //const MODEL_CARD* p = model_dispatcher[base_name];
